@@ -18,24 +18,24 @@
   "license": "MIT",
   "dependencies": {
     "@angular-contrib/core": "0.0.2",
-    "@angular/common": "6.0.0",
-    "@angular/compiler": "6.0.0",
-    "@angular/compiler-cli": "6.0.0",
-    "@angular/core": "6.0.0",
-    "@angular/platform-browser": "6.0.0",
-    "core-js": "2.5.5",
-    "rxjs": "6.1.0",
+    "@angular/common": "6.1.0",
+    "@angular/compiler": "6.1.0",
+    "@angular/compiler-cli": "6.1.0",
+    "@angular/core": "6.1.0",
+    "@angular/platform-browser": "6.1.0",
+    "core-js": "2.5.7",
+    "rxjs": "6.2.2",
     "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "google-closure-compiler": "20180319.0.0",
+    "google-closure-compiler": "20180716.0.1",
     "http-server": "0.11.1",
-    "rimraf": "2.5.4",
-    "rollup": "0.55.0",
-    "rollup-plugin-commonjs": "8.2.6",
-    "rollup-plugin-node-resolve": "3.0.2",
-    "rollup-plugin-uglify": "3.0.0",
-    "typescript": "2.7.2",
-    "yargs": "11.0.0"
+    "rimraf": "2.6.2",
+    "rollup": "0.63.4",
+    "rollup-plugin-commonjs": "9.1.4",
+    "rollup-plugin-node-resolve": "3.3.0",
+    "rollup-plugin-uglify": "4.0.0",
+    "typescript": "2.9.2",
+    "yargs": "12.0.1"
   }
 }
