@@ -9,15 +9,15 @@
     "build-prod": "./node_modules/.bin/dojo build"
   },
   "dependencies": {
-    "@dojo/core": "0.4.0",
-    "@dojo/has": "0.1.2",
-    "@dojo/i18n": "0.5.1",
-    "@dojo/shim": "0.2.6",
-    "@dojo/widget-core": "0.9.6"
+    "@dojo/core": "2.0.0",
+    "@dojo/has": "2.0.0",
+    "@dojo/i18n": "2.0.0",
+    "@dojo/shim": "2.0.0",
+    "@dojo/widget-core": "2.0.7"
   },
   "devDependencies": {
-    "@dojo/cli": "0.6.1",
-    "@dojo/cli-build-app": "0.3.1",
-    "typescript": "2.6.1"
+    "@dojo/cli": "3.0.0",
+    "@dojo/cli-build-app": "3.0.0",
+    "typescript": "2.9.2"
   }
 }
