@@ -20,15 +20,16 @@
     "url": "https://github.com/krausest/js-framework-benchmark.git"
   },
   "devDependencies": {
-    "babel-preset-es2015": "6.18.0",
-    "babel-preset-react": "6.16.0",
-    "webpack": "1.14.0"
+    "babel-preset-es2015": "6.24.1",
+    "babel-preset-react": "6.24.1",
+    "webpack": "4.16.3",
+    "webpack-cli": "^3.1.0"
   },
   "dependencies": {
-    "babel-core": "6.21.0",
-    "babel-loader": "6.2.10",
+    "babel-core": "6.26.3",
+    "babel-loader": "7.1.5",
     "global": "^4.3.2",
     "jsx-loader": "0.13.2",
-    "nervjs": "^1.2.8"
+    "nervjs": "^1.3.2"
   }
 }
