@@ -10,17 +10,18 @@
     "build-prod": "webpack -p"
   },
   "devDependencies": {
-    "babel-core": "6.26.0",
-    "babel-loader": "7.1.2",
+    "babel-core": "6.26.3",
+    "babel-loader": "7.1.5",
     "babel-plugin-transform-react-jsx": "6.24.1",
     "babel-preset-es2015": "6.24.1",
     "babel-preset-stage-0": "6.24.1",
-    "webpack": "3.8.1"
+    "webpack": "4.16.3",
+    "webpack-cli": "^3.1.0"
   },
   "dependencies": {
-    "@cycle/dom": "19.3.0",
+    "@cycle/dom": "20.4.0",
     "@cycle/xstream-run": "4.2.0",
     "snabbdom-jsx": "0.4.2",
-    "xstream": "11.0.0"
+    "xstream": "11.7.0"
   }
 }
