@@ -10,14 +10,15 @@
     "build-prod": "webpack -p"
   },
   "devDependencies": {
-    "babel-core": "6.24.1",
-    "babel-loader": "7.0.0",
+    "babel-core": "6.26.3",
+    "babel-loader": "7.1.5",
     "babel-preset-es2015": "6.24.1",
     "babel-preset-react": "6.24.1",
-    "webpack": "2.5.1",
-    "jsx-loader": "0.13.2"
+    "jsx-loader": "0.13.2",
+    "webpack": "4.16.3",
+    "webpack-cli": "^3.1.0"
   },
   "dependencies": {
-    "etch": "0.12.5"
+    "etch": "0.14.0"
   }
 }
