@@ -24,7 +24,7 @@ var extensions = [
 module.exports = [{
 	cache: cache,
 	module: {
-		loaders: loaders
+		rules: loaders
 	},
 	entry: {
 		main: './src/main.es6.js',
