@@ -3,7 +3,7 @@
 import * as path from "path";
 import babel from "rollup-plugin-babel";
 import resolve from "rollup-plugin-node-resolve";
-import uglify from "rollup-plugin-uglify";
+import { uglify } from "rollup-plugin-uglify";
 import { minify } from "uglify-es";
 
 export default {
