@@ -1,6 +1,6 @@
 import babel from "rollup-plugin-babel"
 import resolve from "rollup-plugin-node-resolve"
-import uglify from "rollup-plugin-uglify"
+import {uglify} from "rollup-plugin-uglify"
 
 export default {
 	plugins: [
